src/main/java/com/tddexample/todo/service/ToDoService.java
package com.tddexample.todo.service;

import com.tddexample.todo.entity.ToDo;
import com.tddexample.todo.repository.ToDoRepository;
import org.springframework.stereotype.Service;

import java.util.ArrayList;
import java.util.List;

@Service
public class ToDoService {

    private ToDoRepository toDoRepository;

    public ToDoService (ToDoRepository toDoRepository) {
        this.toDoRepository = toDoRepository;
    }

    public List<ToDo> findAll() {
        return toDoRepository.findAll();
    }
}
